+++ Hubs/IReservationHubClients.cs
@@ -5,5 +5,5 @@ namespace LaundrySignalR.Hubs;
 public interface IReservationHubClients
 { 
     Task ReservationAdded(ReservationEntry reservationEntry);
-    Task ReservationDeleted(int reservationEntry);
+    Task ReservationDeleted(long reservationEntry);
 }
